chore: finish the TypeScript migration and lock it in
- Convert the 14 remaining tests, test/setup and the vite/vitest configs
- Set allowJs: false and add the config files to the tsconfig include
- Remove the prop-types dependency; nothing imports it any more
- Add plugin:@typescript-eslint/recommended and @typescript-eslint/no-unused-vars,
  and narrow the lint/format globs to ts,tsx
- Fix the four violations the recommended config surfaced: two `catch (err: any)`
  blocks in the auth pages and two ternaries used as statements in
  SettingsProjectPage

src/ is now 100% TypeScript: 0 .js/.jsx files, 432 tests green.
