+++ dataloom-frontend/src/pages/SignUpPage.tsx
@@ -45,10 +45,11 @@ export default function SignUpPage() {
     try {
       await signup(email, password);
       navigate(next || ROUTES.home, { replace: true });
-    } catch (err: any) {
+    } catch (err: unknown) {
       // `detail` is a string for most errors, but an array of objects for
       // FastAPI 422 validation failures — only a string is safe to render.
-      const detail = err?.response?.data?.detail;
+      const detail = (err as { response?: { data?: { detail?: unknown } } } | null)?.response?.data
+        ?.detail;
       showToast(
         typeof detail === "string" ? detail : "Could not create your account. Please try again.",
         "error",
