+++ dataloom-frontend/src/test/FillEmptyForm.test.tsx
@@ -1,11 +1,33 @@
 import { render, screen, waitFor } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
-import { beforeEach, describe, expect, it, vi } from "vitest";
+import { beforeEach, describe, expect, it, vi, type Mock } from "vitest";
 import FillEmptyForm from "../Components/forms/FillEmptyForm";
 import { transformProject } from "../api";
 import { useProjectContext } from "../context/ProjectContext";
 import { useToast } from "../context/ToastContext";
 import usePreviewSave from "../hooks/usePreviewSave";
+
+/** Props accepted by the column-picker doubles stubbed in below. */
+interface StubColumnSelectProps {
+  value: string;
+  onChange: (value: string) => void;
+  options?: string[];
+  placeholder?: string;
+  includeEmptyOption?: boolean;
+  emptyLabel?: string;
+  "data-testid"?: string;
+}
+
+/** Props accepted by the Select double, whose options carry labels. */
+interface StubSelectProps {
+  value: string;
+  onChange: (value: string) => void;
+  options?: { value: string; label: string }[];
+  placeholder?: string;
+  includeEmptyOption?: boolean;
+  emptyLabel?: string;
+  "data-testid"?: string;
+}
 
 vi.mock("../api", () => ({
   transformProject: vi.fn(),
@@ -24,7 +46,7 @@ vi.mock("../hooks/usePreviewSave", () => ({
 }));
 
 vi.mock("../Components/common/ColumnSelect", () => ({
-  default: ({ value, onChange, includeEmptyOption, emptyLabel }) => (
+  default: ({ value, onChange, includeEmptyOption, emptyLabel }: StubColumnSelectProps) => (
     <select aria-label="Column" value={value} onChange={(event) => onChange(event.target.value)}>
       {includeEmptyOption && <option value="">{emptyLabel || "Select column"}</option>}
       <option value="amount">Amount</option>
@@ -34,9 +56,9 @@ vi.mock("../Components/common/ColumnSelect", () => ({
 }));
 
 vi.mock("../Components/common/Select", () => ({
-  default: ({ value, onChange, options }) => (
+  default: ({ value, onChange, options }: StubSelectProps) => (
     <select aria-label="Strategy" value={value} onChange={(event) => onChange(event.target.value)}>
-      {options.map((option) => (
+      {(options ?? []).map((option) => (
         <option key={option.value} value={option.value}>
           {option.label}
         </option>
@@ -44,6 +66,11 @@ vi.mock("../Components/common/Select", () => ({
     </select>
   ),
 }));
+
+const mockTransformProject = transformProject as unknown as Mock;
+const mockUseProjectContext = useProjectContext as unknown as Mock;
+const mockUsePreviewSave = usePreviewSave as unknown as Mock;
+const mockUseToast = useToast as unknown as Mock;
 
 const mockShowToast = vi.fn();
 const mockEnterPreviewMode = vi.fn();
@@ -57,7 +84,7 @@ const renderForm = ({
   columns = ["amount", "created_at"],
   pageSize = 50,
 } = {}) => {
-  useProjectContext.mockReturnValue({
+  mockUseProjectContext.mockReturnValue({
     columns,
     pageSize,
     isPreviewMode,
@@ -65,7 +92,7 @@ const renderForm = ({
     cancelPreview: mockCancelPreview,
   });
 
-  usePreviewSave.mockReturnValue({
+  mockUsePreviewSave.mockReturnValue({
     saving,
     handleSave: mockHandleSave,
   });
@@ -80,11 +107,11 @@ describe("FillEmptyForm", () => {
   beforeEach(() => {
     vi.clearAllMocks();
 
-    useToast.mockReturnValue({
+    mockUseToast.mockReturnValue({
       showToast: mockShowToast,
     });
 
-    transformProject.mockResolvedValue({
+    mockTransformProject.mockResolvedValue({
       columns: ["amount", "created_at"],
       rows: [["100", "2026-07-18"]],
       dtypes: { amount: "integer", created_at: "datetime" },
@@ -133,7 +160,7 @@ describe("FillEmptyForm", () => {
 
     const form = screen.getByRole("button", { name: "Apply" }).closest("form");
     const { fireEvent } = await import("@testing-library/react");
-    fireEvent.submit(form);
+    fireEvent.submit(form!);
 
     await waitFor(() => {
       expect(mockShowToast).toHaveBeenCalledWith(
@@ -194,7 +221,7 @@ describe("FillEmptyForm", () => {
   it("enters preview mode using the transformation response", async () => {
     const user = userEvent.setup();
     const response = { columns: ["amount"], rows: [[1]], dtypes: { amount: "integer" } };
-    transformProject.mockResolvedValue(response);
+    mockTransformProject.mockResolvedValue(response);
 
     renderForm();
     await user.type(screen.getByPlaceholderText("Enter value"), "0");
@@ -228,8 +255,8 @@ describe("FillEmptyForm", () => {
 
   it("shows applying state while the preview request is pending", async () => {
     const user = userEvent.setup();
-    let resolveTransform;
-    transformProject.mockImplementation(
+    let resolveTransform!: (value?: unknown) => void;
+    mockTransformProject.mockImplementation(
       () =>
         new Promise((resolve) => {
           resolveTransform = resolve;
@@ -251,7 +278,7 @@ describe("FillEmptyForm", () => {
 
   it("shows the backend error message when the request fails", async () => {
     const user = userEvent.setup();
-    transformProject.mockRejectedValue({
+    mockTransformProject.mockRejectedValue({
       response: { data: { detail: "Unable to fill empty cells." } },
     });
 
@@ -268,7 +295,7 @@ describe("FillEmptyForm", () => {
 
   it("shows a fallback toast message when the request fails without backend detail", async () => {
     const user = userEvent.setup();
-    transformProject.mockRejectedValue(new Error("Network error"));
+    mockTransformProject.mockRejectedValue(new Error("Network error"));
 
     renderForm();
     await user.type(screen.getByPlaceholderText("Enter value"), "0");
@@ -337,7 +364,7 @@ describe("FillEmptyForm", () => {
       page_size: 10,
     };
 
-    transformProject.mockResolvedValue(response);
+    mockTransformProject.mockResolvedValue(response);
 
     renderForm();
