+++ dataloom-frontend/src/test/setup.ts
@@ -1,0 +1,15 @@
+import "@testing-library/jest-dom";
+
+if (!globalThis.localStorage) {
+  const store = new Map<string, string>();
+
+  Object.defineProperty(globalThis, "localStorage", {
+    value: {
+      getItem: (key: string) => store.get(key) ?? null,
+      setItem: (key: string, value: unknown) => store.set(key, String(value)),
+      removeItem: (key: string) => store.delete(key),
+      clear: () => store.clear(),
+    },
+    configurable: true,
+  });
+}
