+++ dataloom-frontend/tsconfig.json
@@ -17,8 +17,8 @@
     "strict": true,
     "target": "ES2022",
     "jsx": "react-jsx",
-    "allowJs": true,
+    "allowJs": false,
     "noEmit": true
   },
-  "include": ["src"]
+  "include": ["src", "vite.config.ts", "vitest.config.ts"]
 }
